Remove Time#strptime
Was feeling a little uncomfortable with this one. All the tests still
pass, though. So I checked ActiveSupport, and this method is not
overridden there, so IDK if it came from. But anyway, this is already
defined in the Ruby stdlib (at least for the current version). Since I'm
testing on 1.9.3, and it's passing, I'll assume works for all versions
specified in .travis.yml