Refactored the reader method/visibility methods out of Attrio::Attribute.
I used Attrio::AttributeBase as the module name for the shared methods,
but I'm not happy with that name and when I come up with a better name,
I'll rename it.

I left the writer methods in Attrio::Attribute since only Attribute is
going to use those methods.  Factoring those into another module is
something to keep in mind.