Implemented has_element? and find_element for Array, Hash, and Set
There is the potential for adding more advanced find methods.  Don't
think they are needed, but they are there.

for a collection the default value is going to be the return value of
find_element when no element was actually found.  Users can use this to
provide useful defaults such as an appropriate Null class.

Should we provide an :initial_values options to populate the initial
collection?

Next up is to finish the CollectionBuilder and Collection and then to
test integration